Finished breaking out the listeners from main.
