+++ client/listeners/index.js
@@ -1,18 +1,36 @@
+const { attachInitialSocketListeners } = require('./socketListeners')
+
 const attachUrlFormListener = require('./urlSubmitForm')
 const attachAddTrackButtonListener = require('./addTrackButton')
+const attachAudioControlListener = require('./audioControls')
+
+const createTracklistFormListeners = require('./tracklistForm')
 const createResetListeners = require('./resetListeners')
-const attachTracklistFormListener = require('./tracklistForm')
-const { attachInitialSocketListeners } = require('./socketListeners')
+const createTimeCodeListeners = require('./timecodeListeners')
 
 function attachListeners() {
   attachInitialSocketListeners()
   const { attachStartOverBtnListener, attachResetTracklistListener } = createResetListeners()
+  const { attachTracklistFormListener, attachTrackDeleteListener } = createTracklistFormListeners()
+  const {
+    attachLoadTimecodesListener,
+    attachSubmitTimecodesListener,
+    attachCancelTimecodesListener,
+    attachManualTimecodesListener
+  } = createTimeCodeListeners()
+
   return {
-    addTrackButton: attachAddTrackButtonListener(),
+    addTrack: attachAddTrackButtonListener(),
+    deleteTrack: attachTrackDeleteListener(),
     resetTracklistBtn: attachResetTracklistListener(),
     startOverBtn: attachStartOverBtnListener(),
     tracklistForm: attachTracklistFormListener(),
-    urlSubmitForm: attachUrlFormListener()
+    urlSubmitForm: attachUrlFormListener(),
+    audioControls: attachAudioControlListener(),
+    loadTimecodes: attachLoadTimecodesListener(),
+    submitTimecodes: attachSubmitTimecodesListener(),
+    cancelTimecodes: attachCancelTimecodesListener(),
+    manualTimecodes: attachManualTimecodesListener()
   }
 }
